Update to include software awards in Australia
Adding as a suggestion the way we support software for Open Science in Australia. There are two strands: awards for the contribution on new methods, models or ideas as software via scholarly societies, and a larger single national prize for those forms of software that underpin scientific endeavours more broadly.  But, note that none of these is explicitly labelled open science. Many explicitly require submissions of open source software, while all implicitly support open source as a way of producing software in  research, which we believe is under-recognised.